Revert "Revert "split config""
This reverts commit 6d349535b381d26a21c1e0d5c8229a03e3ed7bc1.
