+++ src/routes/borne/config/_external_request_config.svelte
